+++ packages/core/src/actions/setClockSkewAdjustment.js
@@ -1,0 +1,6 @@
+// Currently, this action is for testing purpose only.
+// Thus, action creator is not needed.
+
+const SET_CLOCK_SKEW_ADJUSTMENT = 'WEB_CHAT/SET_CLOCK_SKEW_ADJUSTMENT';
+
+export { SET_CLOCK_SKEW_ADJUSTMENT };