+++ packages/core/src/utils/getTimestamp.js
@@ -1,3 +1,0 @@
-export default function getTimestamp() {
-  return new Date().toISOString();
-}